+++ lib/shelf_packages_handler.dart
@@ -5,32 +5,22 @@
 library shelf_packages_handler;
 
 import 'package:shelf/shelf.dart';
-import 'package:shelf_static/shelf_static.dart';
-import 'package:package_resolver/package_resolver.dart';
-import 'package:path/path.dart' as p;
 
-import 'src/async_handler.dart';
 import 'src/dir_handler.dart';
 import 'src/package_config_handler.dart';
 
 /// A handler that serves the contents of a virtual packages directory.
 ///
 /// This effectively serves `package:${request.url}`. It locates packages using
-/// the package resolution logic defined by [resolver]. If [resolver] isn't
-/// passed, it defaults to the current isolate's package resolution logic.
+/// the mapping defined by [packageMap]. If [packageMap] isn't passed, it
+/// defaults to the current isolate's package resolution logic.
+///
+/// The [packageMap] maps package names to the base uri for resolving
+/// `package:` uris for that package.
 ///
 /// This can only serve assets from `file:` URIs.
-Handler packagesHandler({PackageResolver resolver}) {
-  resolver ??= PackageResolver.current;
-  return AsyncHandler(resolver.packageRoot.then((packageRoot) {
-    if (packageRoot != null) {
-      return createStaticHandler(p.fromUri(packageRoot),
-          serveFilesOutsidePath: true);
-    } else {
-      return PackageConfigHandler(resolver);
-    }
-  }));
-}
+Handler packagesHandler({Map<String, Uri> packageMap}) =>
+    PackageConfigHandler(packageMap: packageMap).handleRequest;
 
 /// A handler that serves virtual `packages/` directories wherever they're
 /// requested.
@@ -40,5 +30,5 @@ Handler packagesHandler({PackageResolver resolver}) {
 ///
 /// This is useful for ensuring that `package:` imports work for all entrypoints
 /// in Dartium.
-Handler packagesDirHandler({PackageResolver resolver}) =>
-    DirHandler('packages', packagesHandler(resolver: resolver));
+Handler packagesDirHandler({Map<String, Uri> packageMap}) =>
+    DirHandler('packages', packagesHandler(packageMap: packageMap));