+++ lib/src/package_config_handler.dart
@@ -3,42 +3,52 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'dart:async';
+import 'dart:isolate';
 
-import 'package:package_resolver/package_resolver.dart';
 import 'package:path/path.dart' as p;
 import 'package:shelf/shelf.dart';
 import 'package:shelf_static/shelf_static.dart';
-
-import 'async_handler.dart';
 
 /// A shelf handler that serves a virtual packages directory based on a package
 /// config.
 class PackageConfigHandler {
   /// The static handlers for serving entries in the package config, indexed by
   /// name.
-  final _packageHandlers = <String, Handler>{};
+  final _packageHandlers = <String, Future<Handler>>{};
 
-  /// The information specifying how to do package resolution.
-  final PackageResolver _resolver;
+  /// Optional, a map of package names to base uri for resolving `package:`
+  /// uris for that package.
+  final Map<String, Uri> _packageMap;
 
-  PackageConfigHandler(this._resolver);
+  PackageConfigHandler({Map<String, Uri> packageMap})
+      : _packageMap = packageMap;
 
   /// The callback for handling a single request.
-  FutureOr<Response> call(Request request) {
+  Future<Response> handleRequest(Request request) async {
     var segments = request.url.pathSegments;
-    return _handlerFor(segments.first)(request.change(path: segments.first));
+    var handler = await _handlerFor(segments.first);
+    return handler(request.change(path: segments.first));
   }
 
-  /// Creates a handler for [package] based on the package map in [_resolver].
-  Handler _handlerFor(String package) {
-    return _packageHandlers.putIfAbsent(package, () {
-      return AsyncHandler(_resolver.urlFor(package).then((url) {
-        var handler = url == null
-            ? (_) => Response.notFound('Package $package not found.')
-            : createStaticHandler(p.fromUri(url), serveFilesOutsidePath: true);
+  /// Creates a handler for [packageName] based on the package map in
+  /// [_packageMap] or the current isolate resolver.
+  Future<Handler> _handlerFor(String packageName) {
+    return _packageHandlers.putIfAbsent(packageName, () async {
+      Uri packageUri;
+      if (_packageMap != null) {
+        packageUri = _packageMap[packageName];
+      } else {
+        var fakeResolvedUri = await Isolate.resolvePackageUri(
+            Uri(scheme: 'package', path: '$packageName/'));
+        packageUri = fakeResolvedUri;
+      }
 
-        return handler;
-      }));
+      var handler = packageUri == null
+          ? (_) => Response.notFound('Package $packageName not found.')
+          : createStaticHandler(p.fromUri(packageUri),
+              serveFilesOutsidePath: true);
+
+      return handler;
     });
   }
 }