+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2.0.0
+
+### Breaking changes
+
+* Dropped the dependency on `package_resolver`.
+  * All `PackageResolver` apis now take a `Map<String, Uri>` of package name
+    to the base uri for resolving `package:` uris for that package.
+  * Named arguments have been renamed from `resolver` to `packageMap`.
+
 ## 1.0.4
 
 * Set max SDK version to `<3.0.0`, and adjust other dependencies.